Rename file (phase II).
