+++ src/commands/containers/files/openContainerFile.ts
